requirements.txt
# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 数据库
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# 数据验证
pydantic>=2.5.3
pydantic-settings>=2.1.0

# 安全
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# 配置
pyyaml>=6.0.1

# HTTP客户端
httpx>=0.26.0
aiohttp>=3.9.1

# JavaScript 引擎
quickjs>=1.19.2

# 工具
python-multipart>=0.0.6
aiofiles>=23.2.1

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0